rename: 메일/소켓 테스트 디렉토리 분리 및 import 정리
