+++ src/main/java/com/icestudyroom_email/domain/email/infrastructure/template/RankingEmailTemplate.java
@@ -1,7 +1,7 @@
 package com.icestudyroom_email.domain.email.infrastructure.template;
 
 import com.icestudyroom_email.domain.email.infrastructure.gmail.dto.EmailRequest;
-import com.icestudyroom_email.domain.rankingContract.RankingEmailEvent;
+import com.icestudyroom_email.domain.rankingContract.email.RankingEmailEvent;
 import com.icestudyroom_email.domain.rankingContract.RankingEventType;
 
 public interface RankingEmailTemplate {
